Release/v1.8.0

+++ api/comments/index.ts
@@ -1,23 +1,27 @@
+import { authorizationHeader } from "../authHeader";
 import { api } from "../myApi";
-import { getToken } from "../getToken";
 import { CommentDeleteProps, CommentProps, CommentUpdateProps } from "./types";
 
 const fetchComments = async ({ quizId }: CommentProps) => {
-  api.defaults.headers.common["jwt"] = getToken();
-  const { data } = await api.get(`/quizzes/${quizId}/comments`);
+  const { data } = await api.get(
+    `/quizzes/${quizId}/comments`,
+    authorizationHeader
+  );
   return data;
 };
 
 const updateComment = async ({ requestData }: CommentUpdateProps) => {
   const { commentBody, quizId } = requestData;
-  api.defaults.headers.common["jwt"] = getToken();
-  return await api.post(`/quizzes/${quizId}/comments`, commentBody);
+
+  return await api.post(
+    `/quizzes/${quizId}/comments`,
+    commentBody,
+    authorizationHeader
+  );
 };
 
 const deleteComment = async ({ commentId }: CommentDeleteProps) => {
-  api.defaults.headers.common["jwt"] = getToken();
-
-  await api.delete(`/comments/${commentId}`);
+  await api.delete(`/comments/${commentId}`, authorizationHeader);
 };
 
 export { fetchComments, updateComment, deleteComment };
